Exported UsageTechnique class to separate file, improved documentation. Refactored scripting attack telemetry sending in pba
