+++ monkey/infection_monkey/post_breach/pba.py
@@ -47,9 +47,20 @@ class PBA(object):
         """
         exec_funct = self._execute_default
         result = exec_funct()
-        if result[1] and isinstance(self.command, list) and len(self.command) > 1:
+        if self.scripts_were_used(result):
             T1064Telem(ScanStatus.USED, "Scripts used to execute %s post breach action." % self.name).send()
         PostBreachTelem(self, result).send()
+
+    def scripts_were_used(self, pba_execution_result):
+        """
+        Determines if scripts were used to execute PBA
+        :param pba_execution_result: result of execution function. e.g. self._execute_default
+        :return: True if scripts were used, False otherwise
+        """
+        pba_execution_succeeded = pba_execution_result[1]
+        if pba_execution_succeeded and isinstance(self.command, list) and len(self.command) > 1:
+            return True
+        return False
 
     def _execute_default(self):
         """
