+++ monkey/monkey_island/cc/services/attack/technique_reports/T1035.py
@@ -1,5 +1,4 @@
-from monkey_island.cc.database import mongo
-from monkey_island.cc.services.attack.technique_reports import UsageTechnique
+from monkey_island.cc.services.attack.technique_reports.usage_technique import UsageTechnique
 
 __author__ = "VakarisZ"
